+++ store/sqlite/src/collection/proc.rs
@@ -70,7 +70,7 @@ impl DbCollection for ProcCollection {
 
         let mut stmt = conn.prepare(sql.as_str()).map_err(map_db_err)?;
         let row = stmt
-            .query_row(&*values.as_params(), |row| data::Proc::from_row(row))
+            .query_row(&*values.as_params(), Self::Item::from_row)
             .map_err(map_db_err)?;
 
         Ok(row)
@@ -133,9 +133,9 @@ impl DbCollection for ProcCollection {
             page_num,
             page_count,
             rows: conn
-                .prepare(&sql.as_str())
+                .prepare(&sql)
                 .map_err(map_db_err)?
-                .query_map(&*values.as_params(), |row| data::Proc::from_row(row))
+                .query_map(&*values.as_params(), Self::Item::from_row)
                 .map_err(map_db_err)?
                 .map(|v| v.unwrap())
                 .collect::<Vec<_>>(),